Move Task class and subclasses to their own package
Renamed the packages in the other java files
